Ci-Cd-30-day-challenge/Python-Project/Day_Five/requirements.txt
lask==2.0.1
pytest==7.2.0

